utilize useNavigate hook to implement routing front-end links to their particular category pages
