+++ src/components/directory-item/directory-item.component.jsx
@@ -1,9 +1,15 @@
+import { useNavigate } from 'react-router-dom';
+
 import { BackgroundImage, DirectoryBody, DirectoryItemContainer } from './directory-item.styles';
 
 const DirectoryItem = ({ category }) => {
-  const { imageUrl, title } = category;
+  const { imageUrl, title, route } = category;
+
+  const navigate = useNavigate();
+  const onNavigateHandler = () => navigate(route);
+
   return (
-    <DirectoryItemContainer>
+    <DirectoryItemContainer onClick={onNavigateHandler}>
       <BackgroundImage imageUrl={imageUrl} />
       <DirectoryBody>
         <h2>{title}</h2>
